# docs/openapi.yml
openapi: 3.0.3
info:
  title: 대타구함 API
  description: |
    알바 근무 대타·교대·급구 조율 서비스 **대타구함**의 RESTful API 명세서입니다.

    ## 인증 방식
    - 휴대폰 번호와 비밀번호로 로그인하면 JWT Access Token을 발급합니다.
    - 이후 모든 요청의 `Authorization` 헤더에 `Bearer {token}`을 포함합니다.
    - 알바생·점장·사장 역할은 계정이 아니라 **매장별**로 정해집니다.
      - 사장: `stores.owner_id`
      - 점장·알바생: `store_members.role` (MANAGER / WORKER)
    - 시스템 관리자는 두지 않습니다. 매장은 초대코드로만 들어가는 **비공개 공간**이고,
      문제가 생기면 사장·점장이 직원을 내보냅니다.

    ## 공통 규칙
    - 날짜/시간 형식: ISO 8601 (예: `2026-09-19T14:00:00`)
    - 페이지네이션: 목록 API에 `page`, `size` 쿼리 파라미터 지원 (선택). 응답은 `PagedResponse` 구조입니다.
    - 시간 겹침: 같은 사람의 근무는 **매장이 달라도** 시간이 겹치면 `409 Conflict`를 반환합니다.
    - 사장은 운영·관리 전용 액터입니다. 근무자로 배정하거나 대타·교대 요청/지원하지 않습니다.
    - 오픈·미들·마감은 기본 시간대입니다. 매장이 정의한 다른 시간대 이름도 허용합니다(최대 10자).
    - 목록의 이름·근무시간·상호 대타 횟수 등은 관계 조회 또는 서버 계산 값이며 ERD에 중복 저장하지 않습니다.

    ## 권한 요약
    | 기능 | 알바생 | 점장 | 사장 |
    |---|:-:|:-:|:-:|
    | 내 근무·대타·교대 요청/지원 | O | O | - |
    | 근무표 관리·승인·직원 관리 | - | O (담당 매장) | O (소유 매장) |
    | 매장 만들기·급구 등록·매장 설정·점장 임명 | - | - | O |
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: 로컬 개발 서버
...
paths:
...
  /stores/{storeId}/shifts:
    get:
      tags: [Shift]
      summary: 매장 근무표 조회
      description: 기간 안의 매장 근무를 조회합니다. 소속 직원 모두 조회할 수 있습니다. (화면 manager/schedule.html)
      parameters:
        - $ref: '#/components/parameters/StoreIdPath'
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2026-09-14'
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2026-09-20'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Shift]
      summary: 근무 등록 (사장·점장)
      description: |
        근무자를 비워두면 빈 근무로 저장됩니다.
        근무자의 다른 근무(다른 매장 포함)와 시간이 겹치면 409를 반환합니다.
      parameters:
        - $ref: '#/components/parameters/StoreIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftWriteRequest'
      responses:
        '201':
          description: 등록 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/TimeConflict'
  /stores/{storeId}/shift-templates:
    get:
      tags: [Shift]
      summary: 매장 시간대 프리셋 조회
      description: |
        이 매장에서 쓰는 시간대입니다. 근무표를 짤 때 이 시간대가 표의 줄이 되고, 칸에 사람을 넣으면 이 시간으로 근무가 만들어집니다.
        이름·시간·개수는 매장이 정합니다. 매장을 만들면 오픈 10:00-16:00, 미들 12:00-18:00, 마감 16:00-22:00 세 개가 기본으로 들어갑니다. (화면 manager/schedule-bulk.html)
      parameters:
        - $ref: '#/components/parameters/StoreIdPath'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftTemplate'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [Shift]
      summary: 매장 시간대 변경 (사장·점장)
      description: |
        보낸 목록으로 통째로 바꿉니다. 시간대를 더 만들거나 지울 수 있습니다.
        이미 저장된 근무의 시간은 바뀌지 않습니다 (shifts에 시간과 시간대 이름을 복사해 두기 때문).
      parameters:
        - $ref: '#/components/parameters/StoreIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ShiftTemplateWriteRequest'
      responses:
        '200':
          description: 변경 완료
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /stores/{storeId}/shifts/bulk:
    post:
      tags: [Shift]
      summary: 근무표 한 번에 저장 (사장·점장)
      description: |
        한 달(또는 정한 기간) 근무표를 통째로 저장합니다. 매장 근무표는 보통 달 단위로 짭니다. (화면 manager/schedule-bulk.html)
        - from ~ to 사이 이 매장 근무를 보낸 목록과 대조합니다. workerId·startAt·endAt·position이 같은 행은 기존 id를 유지하며, 참조되지 않은 행만 삭제·추가합니다. 동일한 값의 행이 여러 개라면 개수까지 대조합니다.
        - 같은 날 같은 시간대에 여러 명이 들어가면 사람 수만큼 근무가 만들어집니다.
        - 한 칸이라도 근무자의 근무 시간이 겹치면(다른 매장 근무, 또는 같은 날 시간이 겹치는 다른 시간대) 아무것도 저장하지 않고 409를 반환하며, 겹친 칸을 conflicts로 알려줍니다.
        - shift_requests.shift_id, application_offer_shifts.shift_id, request_applications.selected_offer_shift_id에 참조된 근무는 요청 상태와 무관하게 보존합니다. 제출 목록에서 해당 행이 빠지거나 값이 달라지면 409 SHIFT_REFERENCED로 전체 저장을 거부합니다. 조용히 남겨 중복 생성하지 않습니다.
        - 초기화는 shifts: []로 저장합니다. 기간 내 참조된 근무가 있으면 동일하게 409로 거부하며 기존 근무표는 그대로 유지합니다.
        - 근무자를 비운 칸은 빈 근무로 저장되어 사장이 급구로 올릴 수 있습니다.
      parameters:
        - $ref: '#/components/parameters/StoreIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftBulkRequest'
      responses:
        '200':
          description: 저장 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftBulkResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 시간 겹침 또는 참조된 근무 변경·삭제로 전체 저장하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftBulkConflict'
...
  /shifts/{shiftId}:
    put:
      tags: [Shift]
      summary: 근무 수정 (사장·점장)
      description: 요청·교대 후보·선택 결과에 참조된 근무의 근무자·시간·포지션 변경은 409 SHIFT_REFERENCED로 거부합니다. 값이 같은 저장은 허용합니다. 요청 확정 전용 트랜잭션에서의 근무자 변경은 이 제한의 대상이 아닙니다.
      parameters:
        - $ref: '#/components/parameters/ShiftIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftWriteRequest'
      responses:
        '200':
          description: 수정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: TIME_CONFLICT (시간 겹침) 또는 SHIFT_REFERENCED (참조된 근무 변경)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Shift]
      summary: 근무 삭제 (사장·점장)
      description: shift_requests.shift_id, application_offer_shifts.shift_id, request_applications.selected_offer_shift_id에 참조되지 않은 근무만 삭제합니다. 취소·완료된 요청도 기록이 남으므로 참조된 근무는 409 SHIFT_REFERENCED로 거부하고 어떤 행도 변경하지 않습니다.
      parameters:
        - $ref: '#/components/parameters/ShiftIdPath'
      responses:
        '204':
          description: 삭제 완료
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 요청·교대 후보·선택 결과에 참조되어 삭제할 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: SHIFT_REFERENCED
                message: 요청 기록에 연결된 근무는 삭제할 수 없습니다.
  /shifts/{shiftId}/candidates:
    get:
      tags: [Shift]
      summary: 대타·교대 부탁할 동료 AI 추천
      description: |
        "한 명에게 부탁" 화면에서 부탁할 동료를 추천합니다. (화면 worker/request-new.html)

        **추천 방식**
        1. 규칙 필터: 범위 안 ACTIVE 직원 중 시간이 겹치지 않는 사람 (교대는 같은 매장만)
        2. 점수: 근무 가능 시간 일치, 이번 주 근무시간이 적은 순, 나를 도와준 횟수
        3. 설명: 판정과 분리된 생성기가 순위별 추천 사유를 한 줄로 생성
           (현재는 규칙 기반 fallback, 외부 AI로 교체 가능)
      parameters:
        - $ref: '#/components/parameters/ShiftIdPath'
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [COVER, EXCHANGE]
        - name: scope
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RequestScope'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
            maximum: 20
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateList'
        '403':
          description: 본인 근무가 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stores/{storeId}/candidates:
    get:
      tags: [Shift]
      summary: 급구 후보 AI 추천 (사장)
      description: |
        급구를 올리기 전에 해당 시간에 일할 수 있는 후보를 추천합니다. (화면 owner/open-shift.html)
        추천 방식은 `GET /shifts/{shiftId}/candidates`와 같고, 최근 급구 참여가 많은 사람은 점수를 낮춥니다.
      parameters:
        - $ref: '#/components/parameters/StoreIdPath'
        - name: startAt
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: '2026-09-25T12:00:00'
        - name: endAt
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: '2026-09-25T20:00:00'
        - name: scope
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/RequestScope'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateList'
        '403':
          description: 사장이 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==========================================
  # [Request] 대타 · 교대 · 급구
  # ==========================================
...
components:
  schemas:
...
    RequestType:
      type: string
      enum: [COVER, EXCHANGE, OPEN_SHIFT]
      description: 'COVER: 대타 / EXCHANGE: 교대 / OPEN_SHIFT: 급구'
...
    RequestScope:
      type: string
      enum: [STORE, OWNER_STORES]
      description: 'STORE: 이 매장 / OWNER_STORES: 같은 사장 매장 전체'
    RequestStatus:
      type: string
      enum: [OPEN, PENDING_APPROVAL, CONFIRMED, CANCELED, EXPIRED]
...
    Shift:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 108
        storeId:
          type: integer
          format: int64
          example: 12
        workerId:
          type: integer
          format: int64
          nullable: true
          description: null이면 빈 근무
        workerName:
          type: string
          nullable: true
          example: 김계원
        startAt:
          type: string
          format: date-time
          example: '2026-09-19T14:00:00'
        endAt:
          type: string
          format: date-time
          example: '2026-09-19T22:00:00'
        position:
          type: string
          example: 마감
        isHelper:
          type: boolean
          description: 근무자가 이 매장 소속이 아닌 헬퍼인지 (서버 계산)
        openRequest:
          $ref: '#/components/schemas/RequestBrief'
...
    ShiftWriteRequest:
      type: object
      required: [startAt, endAt]
      description: 종료는 시작보다 늦어야 합니다. 근무자는 ACTIVE 직원(점장·알바생·허용된 헬퍼)이며 사장은 제외합니다.
      properties:
        workerId:
          type: integer
          format: int64
          nullable: true
          description: 비워두면 빈 근무
        startAt:
          type: string
          format: date-time
          example: '2026-09-18T16:00:00'
        endAt:
          type: string
          format: date-time
          example: '2026-09-18T22:00:00'
        position:
          type: string
          minLength: 1
          maxLength: 10
          description: 매장이 정의한 시간대 이름의 스냅샷. 오픈·미들·마감 외 이름도 허용합니다.
...
    ShiftTemplate:
      type: object
      description: 매장에서 쓰는 시간대 하나
      properties:
        id:
          type: integer
          format: int64
          example: 3
        name:
          type: string
          example: 마감
        startTime:
          type: string
          example: '16:00'
        endTime:
          type: string
          example: '22:00'
        requiredCount:
          type: integer
          description: 이 시간대에 필요한 사람 수
          example: 2
    ShiftTemplateWriteRequest:
      type: object
      required: [name, startTime, endTime]
      description: 매장 내 이름 중복 불가. 종료는 시작보다 늦어야 합니다.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 10
          pattern: '^[^|]+$'
          example: 마감
        startTime:
          type: string
          example: '16:00'
        endTime:
          type: string
          example: '22:00'
        requiredCount:
          type: integer
          minimum: 1
          maximum: 9
          default: 1
          example: 2
    ShiftBulkRequest:
      type: object
      required: [from, to, shifts]
      properties:
        from:
          type: string
          format: date
          example: '2026-10-01'
        to:
          type: string
          format: date
          example: '2026-10-31'
        shifts:
          type: array
          description: 이 기간에 들어갈 근무 전체 (한 달이면 100건 안팎)
          items:
            $ref: '#/components/schemas/ShiftBulkItem'
    ShiftBulkItem:
      type: object
      required: [date, position]
      properties:
        workerId:
          type: integer
          format: int64
          nullable: true
          description: 비워두면 빈 근무 (급구 대상)
        date:
          type: string
          format: date
          example: '2026-10-22'
        position:
          type: string
          description: 시간대 이름. 매장 시간대(shift-templates)의 시간이 그대로 들어갑니다
          minLength: 1
          maxLength: 10
    ShiftBulkResult:
      type: object
      properties:
        created:
          type: integer
          description: 새로 만든 근무 수
          example: 102
        deleted:
          type: integer
          description: 지운 근무 수
          example: 98
        emptyShiftIds:
          type: array
          description: 근무자가 없는 빈 근무 (급구로 올릴 수 있음)
          items:
            type: integer
            format: int64
    ShiftBulkConflict:
      type: object
      description: 시간 겹침 또는 참조된 근무 변경·삭제로 하나도 저장하지 않았을 때
      properties:
        code:
          type: string
          enum: [TIME_CONFLICT, SHIFT_REFERENCED]
          example: TIME_CONFLICT
        message:
          type: string
          example: 겹치는 근무가 있어 저장하지 않았어요.
        conflicts:
          type: array
          description: TIME_CONFLICT의 겹침 목록. SHIFT_REFERENCED일 때는 빈 배열.
          items:
            type: object
            properties:
              workerId:
                type: integer
                format: int64
              workerName:
                type: string
                example: 이진호
              date:
                type: string
                format: date
                example: '2026-10-22'
              storeName:
                type: string
                description: 이미 근무가 잡혀 있는 매장
                example: 건대점
              startAt:
                type: string
                format: date-time
                example: '2026-10-22T18:00:00'
              endAt:
                type: string
                format: date-time
                example: '2026-10-22T22:00:00'
        referencedShiftIds:
          type: array
          description: SHIFT_REFERENCED일 때 변경·삭제가 차단된 기존 근무 id. TIME_CONFLICT일 때는 빈 배열. 기존 FK를 조회한 응답 값이며 별도 DB 컬럼이 아님.
          items:
            type: integer
            format: int64
...
    Candidate:
      type: object
      description: AI 후보 추천 결과 한 명
      properties:
        userId:
          type: integer
          format: int64
        name:
          type: string
          example: 인수연
        storeName:
          type: string
          example: 성수점
        score:
          type: integer
          description: 규칙 기반 점수 (0~100)
          example: 92
        weekHours:
          type: number
          example: 6
        gaveToMe:
          type: integer
          description: 이 사람이 요청자를 도와준 횟수 (대타·교대 추천에서만)
          example: 3
        aiReason:
          type: string
          description: 설명 생성기가 만든 한 줄 추천 사유 (현재는 규칙 기반 fallback)
          example: 이번 주 근무가 가장 적고, 금요일 종일 가능해요
        shifts:
          type: array
          description: 교대 추천일 때 이 사람이 내놓을 수 있는 근무
          items:
            $ref: '#/components/schemas/Shift'
    CandidateList:
      type: object
      properties:
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'
        excludedCount:
          type: integer
          description: 시간이 겹치거나 근무 불가라서 제외된 인원
          example: 3

    # -------------------------------------------------------
    # 요청 · 지원
    # -------------------------------------------------------
...
    RequestBrief:
      type: object
      nullable: true
      description: 근무에 걸린 진행 중 요청 요약 (없으면 null)
      properties:
        requestId:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/RequestType'
        status:
          $ref: '#/components/schemas/RequestStatus'
        applicantCount:
          type: integer
          example: 3
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP 상태 코드
          example: 409
        errorCode:
          type: string
          description: 화면에서 문구를 고르기 위한 세부 코드
          example: TIME_CONFLICT
        message:
          type: string
          example: 이진호님은 그 시간에 건대점 근무가 있어요.
  responses:
    BadRequest:
      description: 잘못된 요청 (필수 값 누락, 형식 오류 등)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: 권한 없음 (다른 매장이거나 역할이 맞지 않음)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 요청한 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TimeConflict:
      description: 근무자의 다른 근무(다른 매장 포함)와 시간이 겹침
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  parameters:
...
    StoreIdPath:
      name: storeId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    ShiftIdPath:
      name: shiftId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...